refactor: 도메인에서 주로 쓰이는 VO domain 패키지로 이동
