+++ src/main/java/chess/dto/MoveRequest.java
@@ -1,7 +1,7 @@
 package chess.dto;
 
+import chess.model.Position;
 import chess.vo.Command;
-import chess.vo.Position;
 
 public class MoveRequest implements Request {
