+++ src/main/java/chess/dto/NotMoveRequest.java
@@ -1,7 +1,7 @@
 package chess.dto;
 
+import chess.model.Position;
 import chess.vo.Command;
-import chess.vo.Position;
 
 public class NotMoveRequest implements Request {
