+++ src/main/java/chess/model/PieceColor.java
@@ -1,4 +1,4 @@
-package chess.vo;
+package chess.model;
 
 public enum PieceColor {
     BLACK, EMPTY, WHITE
