+++ src/main/java/chess/model/Position.java
@@ -1,4 +1,4 @@
-package chess.vo;
+package chess.model;
 
 import static java.lang.String.*;
 import static java.util.stream.Collectors.*;
