+++ src/main/java/chess/model/Rank.java
@@ -1,4 +1,4 @@
-package chess.vo;
+package chess.model;
 
 import static java.util.stream.Collectors.*;
