+++ src/main/java/chess/model/boardinitializer/BoardInitializer.java
@@ -2,8 +2,8 @@ package chess.model.boardinitializer;
 
 import java.util.Map;
 
+import chess.model.Position;
 import chess.model.piece.Piece;
-import chess.vo.Position;
 
 public interface BoardInitializer {
     Map<Position, Piece> apply();
