+++ src/test/java/chess/piece/QueenTest.java
@@ -6,13 +6,13 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 
+import chess.model.File;
+import chess.model.MoveType;
+import chess.model.Path;
+import chess.model.PieceColor;
+import chess.model.Position;
+import chess.model.Rank;
 import chess.model.piece.Queen;
-import chess.vo.File;
-import chess.vo.MoveType;
-import chess.vo.Path;
-import chess.vo.PieceColor;
-import chess.vo.Position;
-import chess.vo.Rank;
 
 public class QueenTest {
